업체 상세정보 화면 내 견적 목록 추가, UI작업 및 페이지네이션 필요

+++ src/test/java/com/glkids/procurehub/repository/QuotationRepositoryTest.java
@@ -35,7 +35,7 @@ public class QuotationRepositoryTest {
                 if (array[1] instanceof QuotationMtrl quotationMtrl) {
                     System.out.println(quotationMtrl.getQtmtno());
                 }
-                if (array[2] instanceof Integer agreementCount) {
+                if (array[2] instanceof Long agreementCount) {
                     System.out.println(agreementCount);
                 }
                 System.out.println();
